requirements.txt
streamlit
google-generativeai
faiss-cpu
numpy
requests
python-dotenv
